refactor: move board to atomic
