+++ src/component/molecules/BoardFooter.tsx
