+++ src/stories/board/BoardFooter.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import BoardFooter from '../../components/board/BoardFooter';
+import BoardFooter from '../../component/molecules/BoardFooter';
 
 export default {
     title: 'components/Board/sub-components',
     component: BoardFooter,
 } as ComponentMeta<typeof BoardFooter>;
 
-const Template: ComponentStory<typeof BoardFooter> = (args) => <BoardFooter {...args}/>;
+const Template: ComponentStory<typeof BoardFooter> = (args) => <BoardFooter {...args} />;
 
 export const Footer = Template.bind({});
 Footer.args = {
