+++ src/stories/board/BoardHeader.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
-import BoardHeader from '../../components/board/BoardHeader';
+import BoardHeader from '../../component/molecules/BoardHeader';
 
 export default {
     title: 'components/Board/sub-components',
     component: BoardHeader,
 } as ComponentMeta<typeof BoardHeader>;
 
-const Template: ComponentStory<typeof BoardHeader> = (args) => <BoardHeader {...args}/>;
+const Template: ComponentStory<typeof BoardHeader> = (args) => <BoardHeader {...args} />;
 
 export const Header = Template.bind({});
 Header.args = {
